+++ KalmanFilter.cpp
@@ -7,20 +7,20 @@ KalmanFilter::KalmanFilter(const cv::Point2i &startPoint){
 													 0, 0, 0, 1);
 
 	
-	measurement(0) = startPoint.x;
-	measurement(1) = startPoint.y;
-	estimated(0) = startPoint.x;
-	estimated(1) = startPoint.y;
+	measurement(0) = (float)(startPoint.x);
+	measurement(1) = (float)(startPoint.y);
+	estimated(0) = (float)(startPoint.x);
+	estimated(1) = (float)(startPoint.y);
 
-	KF.statePre.at<float>(0) = startPoint.x;
-	KF.statePre.at<float>(1) = startPoint.y;
-	KF.statePre.at<float>(2) = startPoint.x;
-	KF.statePre.at<float>(3) = startPoint.y;
+	KF.statePre.at<float>(0) = (float)(startPoint.x);
+	KF.statePre.at<float>(1) = (float)(startPoint.y);
+	KF.statePre.at<float>(2) = (float)(startPoint.x);
+	KF.statePre.at<float>(3) = (float)(startPoint.y);
 
-	KF.statePost.at<float>(0) = startPoint.x;
-	KF.statePost.at<float>(1) = startPoint.y;
-	KF.statePost.at<float>(2) = startPoint.x;
-	KF.statePost.at<float>(3) = startPoint.y;
+	KF.statePost.at<float>(0) = (float)(startPoint.x);
+	KF.statePost.at<float>(1) = (float)(startPoint.y);
+	KF.statePost.at<float>(2) = (float)(startPoint.x);
+	KF.statePost.at<float>(3) = (float)(startPoint.y);
 
 
 	setIdentity(KF.measurementMatrix);
@@ -34,36 +34,36 @@ cv::Point2i KalmanFilter::doFiltering(const cv::Point2i &point) {
 	getPrediction();
 
 	//Get point
-	measurement(0) = point.x;
-	measurement(1) = point.y;
+	measurement(0) = (float)(point.x);
+	measurement(1) = (float)(point.y);
 
 	//The update phase
 	estimated = KF.correct(measurement);
-	cv::Point statePt(estimated.at<float>(0), estimated.at<float>(1));
+	cv::Point statePt((int)(estimated.at<float>(0)), (int)(estimated.at<float>(1)));
 	return statePt;
 }
 
 cv::Point2i KalmanFilter::getPrediction() {
 	cv::Mat prediction = KF.predict();
-	cv::Point predictPt(prediction.at<float>(0), prediction.at<float>(1));
+	cv::Point predictPt((int)(prediction.at<float>(0)), (int)(prediction.at<float>(1)));
 	return predictPt;
 }
 
 
 void KalmanFilter::reset(const cv::Point2i &startPoint){
-	measurement(0) = startPoint.x;
-	measurement(1) = startPoint.y;
-	estimated(0) = startPoint.x;
-	estimated(1) = startPoint.y;
+	measurement(0) = (float)(startPoint.x);
+	measurement(1) = (float)(startPoint.y);
+	estimated(0) = (float)(startPoint.x);
+	estimated(1) = (float)(startPoint.y);
 
-	KF.statePre.at<float>(0) = startPoint.x;
-	KF.statePre.at<float>(1) = startPoint.y;
-	KF.statePre.at<float>(2) = startPoint.x;
-	KF.statePre.at<float>(3) = startPoint.y;
+	KF.statePre.at<float>(0) = (float)(startPoint.x);
+	KF.statePre.at<float>(1) = (float)(startPoint.y);
+	KF.statePre.at<float>(2) = (float)(startPoint.x);
+	KF.statePre.at<float>(3) = (float)(startPoint.y);
 
-	KF.statePost.at<float>(0) = startPoint.x;
-	KF.statePost.at<float>(1) = startPoint.y;
-	KF.statePost.at<float>(2) = startPoint.x;
-	KF.statePost.at<float>(3) = startPoint.y;
+	KF.statePost.at<float>(0) = (float)(startPoint.x);
+	KF.statePost.at<float>(1) = (float)(startPoint.y);
+	KF.statePost.at<float>(2) = (float)(startPoint.x);
+	KF.statePost.at<float>(3) = (float)(startPoint.y);
 }
